pyproject.toml
[project]
name = "weavster"
dynamic = ["version"]
description = "Weavster is a cloud-native integration platform that brings declarative simplicity to data pipelines of all sizes."
authors = [{ name = "Gregory Hunt", email = "huntg1@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
dependencies = [
    "typer >= 0.16.0",
    "fastapi >= 0.116.1",
    "granian >= 2.5.0",
    "trio >= 0.30.0",
    "click != 8.2.2 ",
    "pydantic >= 2.0.0",
    "pyyaml >= 6.0.0",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://weavster-dev.github.io/weavster/"
Repository = "https://github.com/weavster-dev/weavster"
Documentation = "https://weavster-dev.github.io/weavster/"

[project.scripts]
weavster = "weavster.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "ty>=0.0.1a16",
    "pyright>=1.1.390",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "httpx>=0.28.0",
    "pytest-trio>=0.8.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
fallback_version = "0.0.0"

[tool.hatch.build.targets.wheel]
packages = ["src/weavster"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.9"

[tool.ty.src]
exclude = [
    "**/test_*.py"
]

[tool.pyright]
pythonVersion = "3.9"
venvPath = "."
venv = ".venv"

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"**/test_*.py" = ["S101"]

[tool.ruff.format]
preview = true

[tool.deptry.per_rule_ignores]
DEP002 = ["granian", "click"]
DEP004 = ["pytest"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/weavster/server/__init__.py",
]
